Proto,tx_rollup: Introduce an [apply] function for tx rollup layer2
The [apply] function of the layer2 is built thanks to a series of
functors that takes as argument a so-called [STORAGE]. A [STORAGE] is
a simple untyped key-value store. By untyped, we mean that both keys
and values are simple bytes.

On top of this storage, a type-safe “context” (akin to the context of
Tezos) is bulit, and the [apply] function makes use of this
context.

Co-author-by: Marigold <contact@marigold.dev>
Co-author-by: Nomadic Labs <contact@nomadic-labs.com>
Co-author-by: Oxhead Alpha <info@oxheadalpha.com>