Proto: access a contract's stake through an explicit function
Replace 'balance' by 'stake', even though at this point those
are the same for implicit contracts that are not delegates.
This is necessary so that when the stake includes rollup bonds,
it is automatically take into account in the calculation of
baking rights.

Here we have to pass a function as a parameter, to avoid the
circular dependency that would result from calling directly
'Contract_storage.stake'.