Env,proto: Expose a dummy WASM [step] function to the protocol
In this patch, we enrich the V6 protocol with a functor to derive a
WASM step function.

The functor is parameterized with an Irmin tree, so that it can easily
be used with different context implementation.  This is necessary as
long as (1) the functor of the WASM PVM is parameterized by a tree,
and (2) the protocol only knows about one Irmin tree (32-arity,
specific to Tezos needs).