Proto,SCORU: Fix bug in rollup node
When the queue of heads to be processed is not empty and the rollup
node is terminated, the heads in the queue are never processed when
the rollup node is restarted.

This commit fixes the issue by remembering the processed heads and
to include unprocessed heads in the intermediate heads to catch
up with the mainnet heads.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>