Proto,sc_rollup: Ask for an origination proof, compute a genesis hash
This origination proof is expected to cover the specialization of a
PVM with a specific boot sector.  Using the helpers we introduce in
this patch, we add tests to check the correctness of the
[initial_state_hash] constant provided by each PVM.

This patch is larger than one could expect because the wrapped proof
definition in `Alpha_context' was breaking the abstraction barrier.

The so-called genesis commitment is still not computed.

Co-authored-by: Sylvain Ribstein <sylvain.ribstein@gmail.com>