Proto: Introduce a new invariant on Michelson value sizes
To make a rationale usage of memory, we maintain the following
invariant:

No Michelson instruction should be able to push a value on the
stack which cannot be serialized to Micheline without going
beyond the hard limit of gas (for one operation).

We implement this invariant by adding a threshold to the gas
model of every instruction that can generate a value of dynamic
length.

This commit only implements the case for concat_string/bytes to
validate the failing tests.

In subsequent commits, we will introduce one failing test for each
instructions constructing dynamically sized values followed by a
commit that fixes the failing test.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>