Proto,SCORU: Optimize valid_back_path
The check done on each element of the path of a skip list to ensure
that it is valid (i.e., connecting source to target) and minimal
(i.e., there is no other valid path that is smaller than this one)
was suboptimal.

Indeed, it was check that the pointer under focus was one of the
backpointer of the current cell and that this pointer was indeed the
minimal pointer (computed by `best_skip`). The first check is
redundant as the minimal pointer is necessarily one of the
backpointer.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>