WASM: Make 'From_block' resolution its own tick
Prior to this patch, once the block behind a 'From_block' admin_instr
was resolved, the tick continued.  This might be fine, but it is not
really in line with “a tick is about one thing.”

For consistency, we make 'Eval.step' immediately return after this
step, rather than calling 'step_resolved'.