WASM: Provide a binary to trace tick size of kernel execution
The WASM interpreter handles the memory and table related instruction
using rewriting rules. For instance, MemoryInit is rewritten into six
instructions to call push values on the stack to call Store, then
MemoryInit again.

This is highly inefficient, and can easily explain the big number of
ticks we are witnessing when executing real world WASM code.

This patch inline the effects on MemomyInit in the interpretation of
this instruction, instead of rewriting it. And we keep the logic that
MemoryInit “calls” itself byte per byte, so there are still rooms from
improvement. But the immediate consequence of this patch is that
computation.wasm takes 18K ticks to execute, instead of 36K. A
significant reduction.