WASM: Filter exports when run by the PVM
A WASM program can export memories, functions, tables, globals... that
other modules can use. These exported resources are constructed during
the init phase, in such a way than when the module modifies a memory
that is exported for instance, the “exported” version of this memory
(which, again, might be used by another module) is modified
accordingly. This is done, as often in the WASM interpreter, through
reference sharing.

And this sharing is broken (again) by the decode-encode loop used by
the PVM. As soon as the references are created, they are untied by the
tree-encoding, which means that further modification of the memory (or
table, etc.) of a module do not modify its “extern” counterpart.

This is not a problem in the WASM PVM, because we do not have support
for multi-modules program. But it does break the correspondence
between 1 N-ticks and N 1-tick executions.