WASM: Fix locals management in the stack frames
During the evaluation of a WASM program, the interpreter maintains a
stack of stack frames (on per function call), plus a stack of labels
within each frame. A stack frame and its labels gets to modify a stack
of values and instructions, encoded in values of the type [frame].

This works well enough on 1 N-tick executions thanks to aliasing and
shared mutability. But the latter are totally lost when considering N
1-tick executions. This leads to hash inconsistencies, and potentially
bugs in the N 1-tick executions.

This patch addresses this issue, through a collection of changes that
together align the behavior of both execution models.

  - Don't duplicate frame inst in [frame_stack] and [label]
  - Store locals as values, not reference of values, turn the locals
    field of the frame into a mutable field
  - Break any aliasing explicitly to avoid hash inconsistencies
    between the two execution models[*]

[*] These inconsistencies would not lead to execution result
    divergence, but would prevent using the 1 N-tick execution as a
    way to speed up the execution of a WASM program using the PVM.