WASM/PVM: add representation of input from L1
This representation can be thought of a half-deserialized version of
[Sc_rollup_inbox_message_repr.t]. This type cannot be used as is from the
protocol, since `lib_scoru_wasm` lives under the environment barrier, and the
protocol is above. Moreover, some of its values are serialized types that does
not exist outside of the protocol (Contract and rollups hash for example), which
makes it difficult to write encoders and decoders for it.

We take the approach to simply read the first two bytes of the input to
discriminate the input and leaves deserialization to the kernel.