Proto,SCORU: Require conflicting commitments hashes to start game
Since get_conflict_point is not gas friendly, we stop using it
in the protocol when we start a game. This computation must be
done offchain.

This commit is a bit large but it simply amounts to push the
commitment hashes provided by the initial refutation.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>
(cherry picked from commit 87a2f15fe832b991ef416993fbf8b21157dca581)