Manifest: add tezos-bls12-381-polynomial-internal
The vendored version of tezos-bls12-381-polynomial is called temporarily
tezos-bls12-381-polynomial-internal and is used only in DAL related projects.
tezos-bls12-381-polynomial is used in the dependency tezos-plonk and
tezos-plompiler, and it introduces therefore cycle dependencies.

tezos-plompiler, tezos-plonk, tezos-aplonk, tezos-epoxy-tx and
tezos-distributed-plonk will be moved later in tezos/tezos. When this is done,
tezos-bls12-381-polynomial-internal can be renamed into
tezos-bls12-381-polynomial.

This commit also renames the module Bls12_381_polynomial in
Tezos_bls12_381_polynomial_internal to call the correct library.