squash! SCORU/Node: worker for playing a single refutation game
For review:

The module [Refutation_player] defines a worker group that will be
handled by the refutation coordinator. A new refutation player will
be spawned for each new conflict that the node has to defend.

A [Refutation_player] can be initialized on a fresh conflict, and
therefore tasked with playing the opening move, or for an ongoing game.

This commit will be squashed.