WIP: Proto/Michelson: remove legacy annotations on option
Field annotations used to be allowed on the `option` Michelson
type. This feature was buggy and got removed in Babylon but it is
still supported in legacy mode. MR !3730 rewrote all annotations on
options in existing scripts during the Jakarta activation so we can
safely remove this legacy code now.