CI: refactor [scripts/ci/doc_publish.sh]
Build the documentation unconditionally. This makes the job easier to
test in artificial [master_branch] pipelines.

Move the last conditional checking whether there is any changes to
commit inside the conditional that checks whether we are running on
master. Indeed, there is no point of pushing unless we've checked out
the documentation repository.