Tests/Proto/Helpers: fix get_endorsing_power_for_delegate
This helper function currently stops at the first item in the list of
endorsers whose delegate is the requested one. However, When several
levels are requested, the same endorser may appear several times in
the list.

This commit modifies the helper function to make it count all the
slots of the endorser list which correspond to the requested address
instead of stoping at the first match.