PlonK/Kzg & Cq: unify commitment in Utils
This commit also changes the pippenger function used : before it was Srs_gX.pippenger & now it’s GX.pippenger_with_affine_array. For Cq, SRS is already given as a G1.array in public parameters but in KZG, it’s still as Srs_g1.t and the conversion is done each time we commit (this is still a ×2 gain on each commit). It would be nice to have the SRS as G1.t array in the public parameters to avoid this conversion, & maybe consider to directly have the affine_array