Tests/Migration: bake more blocks after migration test
At the end of the middle-of-cycle migration tests, 5 blocks are
baked. This is not always enough to test that the migration resulted
in a working state; in particular it may be too short to trigger a
cycle change because cycle length is 8 in these tests.

In this commit, we change the number of baked blocks after the
migration to be twice preserved_cycles, which is enough not only to
trigger a cycle change but also to apply the stake distribution
resulting from this cycle change.