MIR: Add high-level storage operations
Problem: the durable storage API provided by `tezos_smart_rollup` set of
libraries is inconvenient in practice. Just to check the storage for a
value I have to:
* Either check for the key presence first, or sort out the errors.
* Do something with the errors, some of which are impossible, and most
  of which are not recoverable.
* Overall, not to forget that the key can be not present in the storage,
  because the natural wish is to just to propagate the error.

Solution: start a layer on top of that API that:
* Hides operations for working with slices for now, we only read or
  write entire value at a time.
* Sorts out the errors: key absence is reported via `None`, all
  de-facto unchecked errors are turned into panics. Examples of such
  errors: key is too large, some generic error at the host side.
* Very clearly returns the important information, like whether `get`
  found no specified key, or whether `move` has actually moved anything.

The point about panics is arguable, we may want to consider going with
explicit `InternalError` put in `Result`, but I don't see it being
beneficial:
* Those errors are mostly always a programmer mistake;
* There is no way to recover from them;
* They are trully internal, we never want to report them to the user.

This commits yet only adds two functions necessary for us - reading and
writing.