Kaitai: Fix list translation
In kaitai `repeat` spec of an attribute is used to model list/array
types. In case of `Dynamic_size` constructor wrapping a list
attribute, we also add a `size` spec to it, e.g. `size_of_list`.

The problem is that `size` spec can only be used with `bytes`
or user defined types. To make this work, we use `user_type_of_attrs`
instead of `redirect_if_many` helper, so that we always guaranteei a
user type.

Alternativel we could nest logic for list handling inside
`(Dynamic_size(List...))` case as well so that we use
`user_type_of_attrs` there, and keep `redirect_if_many`
for non wrapped list types.