MIR: Catch panics in tzt-runner
Problem: if a normal error occurs, we continue processing other tests.
But not in case of a panic, which are still possible at least because we
have plenty of `todo!(...)`s.

I want to run things like `tzt-runner pack*` and not worry about some
tests being not checked.

Solution: catch panics in the loop that goes over files.