EVM/Exec: prevent FatalError when intermediate transaction runs out of gas
An intermediate call can have a gas value to the sub-context where the
execution is done. This does not mean that an OutOfGas error should end
the entire transaction as there could still be other execution to proceed
and still enough gas at one level above the depth where the OOG error
happens.