CI: refactor, extract [changes:] from [oc.unified_coverage] and align
The intent of the rule on the [unified_coverage] is to run it when any
of its dependencies run, namely tezt and unit jobs. The latter jobs
have the [.rules__octez_changes] template. However, the [changes:]
list of [unified_coverage] does not align.

We need to add an extra condition to stop marge-bot from running this
job. For this reason we need to create a new template.