Kernel/EVM: ignore priority fee
priority fee is a mechanism to allow the user to 'tip' the producer of a block.

However, this doesn't apply on etherlink. The producer/sequencer is already compensated via the data availability & gas fees.
Therefore, we choose to simplify by ignoring priority fee for the purpose of gas price calculation.