CI: Use [scripts/ci/sccache-start.sh] to work around sccache issue
With this change, we start up [sccache] explicitly in the beginning of
the job. We attempt to start sccache 4 times (note: each attempt takes
at most 10 seconds before timing out). Once started, we ensure it
stays awake by disabling the [SCCACHE_IDLE_TIMEOUT]. For good measure,
we also stop it explictly at the end of the script, which has the
positive side-effect of printing statistics.